+++ spring-integration-core/src/main/java/org/springframework/integration/support/management/LifecycleMessageHandlerMetrics.java
@@ -42,6 +42,10 @@ public class LifecycleMessageHandlerMetrics implements MessageHandlerMetrics, Li
 		this.delegate = delegate;
 	}
 
+	public MessageHandlerMetrics getDelegate() {
+		return this.delegate;
+	}
+
 	@SuppressWarnings("unchecked")
 	@Override
 	public void configureMetrics(AbstractMessageHandlerMetrics metrics) {
